requirements.txt
Pillow
streamlit
altair<5